+++ src/common/user-status.enum.ts
@@ -1,5 +1,0 @@
-export enum UserStatus {
-  ACTIVE = 'active',
-  INACTIVE = 'inactive',
-  BANNED = 'banned',
-}