+++ src/modules/auth/strategies/jwt.strategy.ts
@@ -1,0 +1,24 @@
+import { Injectable, UnauthorizedException } from '@nestjs/common';
+import { ConfigService } from '@nestjs/config';
+import { PassportStrategy } from '@nestjs/passport';
+import { ExtractJwt, Strategy } from 'passport-jwt';
+import { JwtPayload } from '../interfaces/jwt-payload.interface';
+
+@Injectable()
+export class JwtStrategy extends PassportStrategy(Strategy) {
+  constructor(configService: ConfigService) {
+    super({
+      jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
+      ignoreExpiration: false, // Reject expired token
+      secretOrKey: configService.get<string>('jwt.accessSecret')!,
+    });
+  }
+
+  async validate(payload: JwtPayload) {
+    // NestJS will automatically attach the return value to request.user
+    if (payload.sub == null) {
+      throw new UnauthorizedException('User is not found');
+    }
+    return { id: payload.sub, email: payload.email, roles: payload.roles };
+  }
+}